+++ src/App.js
@@ -1,18 +1,19 @@
-import React from 'react';
-import { renderRoutes } from 'react-router-config';
-import { Box, Flex, Text } from 'rebass';
-function App({route}) {
-  console.log("ROUTE", route)
+import React from 'react'
+import { renderRoutes } from 'react-router-config'
+import { Box } from 'rebass'
+
+function App({ route }) {
   return (
     <Box
       sx={{
         maxWidth: 512,
         mx: 'auto',
-        px: 3,
-      }}>
-    {renderRoutes(route.routes)}
-  </Box>
-  );
+        px: 3
+      }}
+    >
+      {renderRoutes(route.routes)}
+    </Box>
+  )
 }
 
-export default App;
+export default App
