+++ src/routes/routes.js
@@ -1,35 +1,21 @@
-import App from '../App';
-import CreditCardForm from '../CreditCardForm'
-import Test from '../Test'
+import App from '../App'
+import CreditCardForm from '../components/CreditCardForm/CreditCardForm'
+import Test from '../components/Test/Test'
 
 export const routes = [
   {
     component: App,
-    
-    // @TODO make it works
-    // routes: [
-    //   {
-    //     path: '/test',
-    //     exact:true,
-    //     component: React.lazy(() => import('../Test'))
-    //   },
-    //   {
-    //     path: '/credit-card',
-    //     component: React.lazy(() => import('../CreditCardForm'))
-    //   }
-    // ]
     routes: [
       {
         path: '/test',
         exact: true,
-        component:  Test,
+        component: Test
       },
       {
         path: '/credit-card',
-        exact:true,
-        component: CreditCardForm,
+        exact: true,
+        component: CreditCardForm
       }
     ]
-  },
-  
+  }
 ]
